+++ packages/core/src/nominator.ts
@@ -141,89 +141,78 @@ export default class Nominator {
     this._avgStake = avgStake;
   }
 
-  public async nominate(
-    targets: Types.Stash[],
-    dryRun = false,
-  ): Promise<boolean> {
+  public async nominate(targets: Types.Stash[]): Promise<boolean> {
     const now = new Date().getTime();
 
-    if (dryRun) {
-      logger.info(`DRY RUN - STUBBING TRANSACTIONS`, label);
-      for (const stash of targets) {
-        await queries.setTarget(this.controller, stash, now);
-        await queries.setLastNomination(this.controller, now);
+    const api = this.handler.getApi();
+
+    let tx: SubmittableExtrinsic<"promise">;
+
+    // Start an announcement for a delayed proxy tx
+    if (this._isProxy && this._proxyDelay > 0) {
+      logger.info(
+        `{Nominator::nominate::proxy} starting tx for ${this.address} with proxy delay ${this._proxyDelay} blocks`,
+        label,
+      );
+
+      const innerTx = api.tx.staking.nominate(targets);
+
+      const currentBlock = await api.rpc.chain.getBlock();
+      const { number } = currentBlock.block.header;
+      const callHash = innerTx.method.hash.toString();
+
+      tx = api.tx.proxy.announce(
+        this.controller,
+        blake2AsHex(innerTx.method.toU8a()),
+      );
+      await queries.addDelayedTx(
+        number.toNumber(),
+        this.controller,
+        targets,
+        callHash,
+      );
+
+      try {
+        await tx.signAndSend(this.signer);
+      } catch (e) {
+        logger.error(
+          `{Nominator::nominate} there was an error sending the tx`,
+          label,
+        );
+        logger.error(e);
       }
+    } else if (this._isProxy && this._proxyDelay == 0) {
+      // Start a normal proxy tx call
+      logger.info(
+        `{Nominator::nominate::proxy} starting tx for ${this.address} with proxy delay ${this._proxyDelay} blocks`,
+        label,
+      );
+
+      const innerTx = api.tx.staking.nominate(targets);
+      const callHash = innerTx.method.hash.toString();
+
+      const outerTx = api.tx.proxy.proxy(this.controller, "Staking", innerTx);
+
+      const [didSend, finalizedBlockHash] = await this.sendStakingTx(
+        outerTx,
+        targets,
+      );
+
+      const nominateMsg = `{Nominator::nominate::proxy} non-delay ${this.address} sent tx: ${didSend} finalized in block #${finalizedBlockHash}`;
+      logger.info(nominateMsg, label);
+      if (this.bot) await this.bot.sendMessage(nominateMsg);
     } else {
-      const api = this.handler.getApi();
-
-      let tx: SubmittableExtrinsic<"promise">;
-
-      // Start an announcement for a delayed proxy tx
-      if (this._isProxy && this._proxyDelay > 0) {
-        logger.info(
-          `{Nominator::nominate::proxy} starting tx for ${this.address} with proxy delay ${this._proxyDelay} blocks`,
-          label,
-        );
-
-        const innerTx = api.tx.staking.nominate(targets);
-
-        const currentBlock = await api.rpc.chain.getBlock();
-        const { number } = currentBlock.block.header;
-        const callHash = innerTx.method.hash.toString();
-
-        tx = api.tx.proxy.announce(
-          this.controller,
-          blake2AsHex(innerTx.method.toU8a()),
-        );
-        await queries.addDelayedTx(
-          number.toNumber(),
-          this.controller,
-          targets,
-          callHash,
-        );
-
-        try {
-          await tx.signAndSend(this.signer);
-        } catch (e) {
-          logger.error(
-            `{Nominator::nominate} there was an error sending the tx`,
-            label,
-          );
-          logger.error(e);
-        }
-      } else if (this._isProxy && this._proxyDelay == 0) {
-        // Start a normal proxy tx call
-        logger.info(
-          `{Nominator::nominate::proxy} starting tx for ${this.address} with proxy delay ${this._proxyDelay} blocks`,
-          label,
-        );
-
-        const innerTx = api.tx.staking.nominate(targets);
-        const callHash = innerTx.method.hash.toString();
-
-        const outerTx = api.tx.proxy.proxy(this.controller, "Staking", innerTx);
-
-        const [didSend, finalizedBlockHash] = await this.sendStakingTx(
-          outerTx,
-          targets,
-        );
-
-        const nominateMsg = `{Nominator::nominate::proxy} non-delay ${this.address} sent tx: ${didSend} finalized in block #${finalizedBlockHash}`;
-        logger.info(nominateMsg, label);
-        if (this.bot) await this.bot.sendMessage(nominateMsg);
-      } else {
-        // Do a non-proxy tx
-        logger.info(
-          `(Nominator::nominate) Creating extrinsic Staking::nominate from ${this.address} to targets ${targets} at ${now}`,
-          label,
-        );
-        tx = api.tx.staking.nominate(targets);
-        logger.info(
-          "(Nominator::nominate} Sending extrinsic to network...",
-          label,
-        );
-        await this.sendStakingTx(tx, targets);
-      }
+      // Do a non-proxy tx
+      logger.info(
+        `(Nominator::nominate) Creating extrinsic Staking::nominate from ${this.address} to targets ${targets} at ${now}`,
+        label,
+      );
+      tx = api.tx.staking.nominate(targets);
+      logger.info(
+        "(Nominator::nominate} Sending extrinsic to network...",
+        label,
+      );
+      await this.sendStakingTx(tx, targets);
     }
 
     return true;
